Added the login functionality

+++ studybud/base/views.py
@@ -1,6 +1,7 @@
 from django.db.models import Q
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
@@ -15,6 +16,9 @@ rooms = [
 
 def loginPage(request):
 
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -58,6 +62,7 @@ def room(request, pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def cerateRoom(request):
     form = RoomForm()
 
@@ -70,9 +75,13 @@ def cerateRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -83,8 +92,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
